fix: refactor mdx
