+++ libs/ui/src/lib/catalog-promo/catalog-promo.module.scss
