+++ libs/ui/src/lib/markdown/index.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import ReactMarkdown from 'react-markdown';
-import { Link, LinkProps } from '@digdir/designsystemet-react';
-import { Prism as SyntaxHighlighter, SyntaxHighlighterProps } from 'react-syntax-highlighter';
-import { vscDarkPlus } from 'react-syntax-highlighter/dist/esm/styles/prism';
+import { i18n, type LocaleCodes } from '@fdk-frontend/dictionaries';
+import { mdxComponents } from './components';
 
 export const defaultAllowedElements = [
     'p',
@@ -35,61 +34,24 @@ export const defaultAllowedElements = [
 ];
 
 export type MarkdownProps = {
+    locale?: LocaleCodes;
     className?: string;
     children?: string;
     allowedElements?: string[];
     components?: any;
 };
 
-const Markdown = ({ allowedElements = defaultAllowedElements, components, ...rest }: MarkdownProps) => {
+const Markdown = ({
+    allowedElements = defaultAllowedElements,
+    components = {},
+    locale = i18n.defaultLocale,
+    ...rest
+}: MarkdownProps) => {
     return (
         <ReactMarkdown
             allowedElements={allowedElements}
             components={{
-                a: (props: LinkProps) => <Link {...props} />,
-                code: ({ className, ...props }: SyntaxHighlighterProps) => {
-                    // return <code {...props} />;
-                    const match = /language-(\w+)/.exec(className || '');
-                    return match ? (
-                        <SyntaxHighlighter
-                            style={vscDarkPlus as any}
-                            language={match[1]}
-                            PreTag='div'
-                            {...props}
-                        />
-                    ) : (
-                        <SyntaxHighlighter
-                            style={vscDarkPlus as any}
-                            PreTag='div'
-                            {...props}
-                        />
-                    );
-                },
-                p: ({ children }: { children: React.ReactNode }) => {
-                    return (
-                        <p>
-                            {React.Children.map(children, (child) => {
-                                if (typeof child === 'string') {
-                                    return child.split(/(https?:\/\/[^\s]+)/g).map((part, i) =>
-                                        /^https?:\/\//.test(part) ? (
-                                            <Link
-                                                key={i}
-                                                href={part}
-                                                target='_blank'
-                                                rel='noopener noreferrer'
-                                            >
-                                                {part}
-                                            </Link>
-                                        ) : (
-                                            part
-                                        ),
-                                    );
-                                }
-                                return child;
-                            })}
-                        </p>
-                    );
-                },
+                ...mdxComponents({ locale }),
                 ...components,
             }}
             {...rest}
@@ -98,3 +60,4 @@ const Markdown = ({ allowedElements = defaultAllowedElements, components, ...res
 };
 
 export default Markdown;
+export * from './components';
